refactor: getOrders.success 리듀서에서 데이터 전처리 하고, 스토어에 저장하는 방식에서 -> ordersService에 전처리 로직을 담는 방식으로 수정
  - refactor: GetOrdersSuccessPayload interface의 orders의 타입 OrderItemFromServer[]에서 OrderItem[] 으로 변경

  - refactor: getOrdersSaga에서 call(ordersService.getOrders)에서 받아오는 데이터 타입 OrderItemFromServer[]에서 OrderItem[] 으로 수정

  - refactor: 데이터 전처리하는 함수(createRefinedOrders,divideProductAndQuantity) orderSlice 아래 위치시키는 방식에서 ordersService 아래 두는 방식으로 수정
