+++ client/store/modules/orders/saga.ts
@@ -1,12 +1,12 @@
 import { call, put, takeLatest } from '@redux-saga/core/effects';
 
 import ordersService from '../../../apis/orders';
+import { OrderItem } from '../../../shared/types';
 import { ordersActions } from './slice';
-import { OrderItemFromServer } from './types';
 
 function* getOrdersSaga() {
   try {
-    const orders: OrderItemFromServer[] = yield call(ordersService.getOrders);
+    const orders: OrderItem[] = yield call(ordersService.getOrders);
 
     yield put(ordersActions.getOrders.success({ orders }));
   } catch (error) {
